backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
python-dotenv==0.19.2
requests==2.28.2
Pillow==9.3.0
